requirements.txt
selenium
kiteconnect
requests
pyyaml
